Struct splines::BezierCurve

pub struct BezierCurve {
    // some fields omitted
}

Methods

impl BezierCurve

fn create_path(&self, acc: f32) -> Path

fn evaluate(&self) -> Vec<Vec4<f32>>

fn length(&self, acc: f32, pb: &mut PathBuilder) -> f32

Approximates the length of a Bezier curve within a given accuracy.

fn split(&self, t: f32) -> (BezierCurve, BezierCurve)

Splits an existing Bezier curve into two parts at parameter value t.